src/main/java/ru/skypro/kursovaia/service/ExaminerServiceImpl.java
package ru.skypro.kursovaia.service;

import org.springframework.stereotype.Service;
import ru.skypro.kursovaia.exeption.IncorrectQuestionAmountException;
import ru.skypro.kursovaia.model.Question;

import java.util.Collection;
import java.util.HashSet;
import java.util.Set;

@Service
public class ExaminerServiceImpl implements ExaminerService {
    private final QuestionService questionService;

    public ExaminerServiceImpl(QuestionService questionService) {
        this.questionService = questionService;
    }

    @Override
    public Collection<Question> getQuestion(int amount) {
        if (amount <= 0 || amount > questionService.getAll().size()) {
            throw new IncorrectQuestionAmountException();
        }
        Set<Question> questions = new HashSet<>();
        while (questions.size() < amount) {
            questions.add(questionService.getRandomQuestion());
        }
        return null;
    }
}
